Enum pallet_mapped_assets::pallet::Event

pub enum Event<T: Config<I>, I: 'static = ()> {

    Created {
        asset_id: T::AssetId,
        creator: T::AccountId,
        owner: T::AccountId,
    },
    Issued {
        asset_id: T::AssetId,
        owner: T::AccountId,
        total_supply: T::Balance,
    },
    Transferred {
        asset_id: T::AssetId,
        from: T::AccountId,
        to: T::AccountId,
        amount: T::Balance,
    },
    Reserved {
        reserve_id: T::ReserveIdentifier,
        asset_id: T::AssetId,
        who: T::AccountId,
        amount: T::Balance,
    },
    Unreserved {
        reserve_id: T::ReserveIdentifier,
        asset_id: T::AssetId,
        who: T::AccountId,
        amount: T::Balance,
    },
    BurnedReserve {
        reserve_id: T::ReserveIdentifier,
        asset_id: T::AssetId,
        who: T::AccountId,
        amount: T::Balance,
    },
    TransferredReserve {
        reserve_id: T::ReserveIdentifier,
        asset_id: T::AssetId,
        from: T::AccountId,
        to: T::AccountId,
        amount: T::Balance,
    },
    Burned {
        asset_id: T::AssetId,
        owner: T::AccountId,
        balance: T::Balance,
    },
    TeamChanged {
        asset_id: T::AssetId,
        issuer: T::AccountId,
        admin: T::AccountId,
        freezer: T::AccountId,
    },
    OwnerChanged {
        asset_id: T::AssetId,
        owner: T::AccountId,
    },
    Frozen {
        asset_id: T::AssetId,
        who: T::AccountId,
    },
    Thawed {
        asset_id: T::AssetId,
        who: T::AccountId,
    },
    AssetFrozen {
        asset_id: T::AssetId,
    },
    AssetThawed {
        asset_id: T::AssetId,
    },
    AccountsDestroyed {
        asset_id: T::AssetId,
        accounts_destroyed: u32,
        accounts_remaining: u32,
    },
    ApprovalsDestroyed {
        asset_id: T::AssetId,
        approvals_destroyed: u32,
        approvals_remaining: u32,
    },
    DestructionStarted {
        asset_id: T::AssetId,
    },
    Destroyed {
        asset_id: T::AssetId,
    },
    ForceCreated {
        asset_id: T::AssetId,
        owner: T::AccountId,
    },
    MetadataSet {
        asset_id: T::AssetId,
        name: Vec<u8>,
        symbol: Vec<u8>,
        decimals: u8,
        is_frozen: bool,
    },
    MetadataCleared {
        asset_id: T::AssetId,
    },
    ApprovedTransfer {
        asset_id: T::AssetId,
        source: T::AccountId,
        delegate: T::AccountId,
        amount: T::Balance,
    },
    ApprovalCancelled {
        asset_id: T::AssetId,
        owner: T::AccountId,
        delegate: T::AccountId,
    },
    TransferredApproved {
        asset_id: T::AssetId,
        owner: T::AccountId,
        delegate: T::AccountId,
        destination: T::AccountId,
        amount: T::Balance,
    },
    AssetStatusChanged {
        asset_id: T::AssetId,
    },
    // some variants omitted

}

The event emitted by this pallet.

Variants

Created

Fields

asset_id: T::AssetId

creator: T::AccountId

owner: T::AccountId

Some asset class was created.

Issued

Fields

asset_id: T::AssetId

owner: T::AccountId

total_supply: T::Balance

Some assets were issued.

Transferred

Fields

asset_id: T::AssetId

from: T::AccountId

to: T::AccountId

amount: T::Balance

Some assets were transferred.

Reserved

Fields

reserve_id: T::ReserveIdentifier

asset_id: T::AssetId

who: T::AccountId

amount: T::Balance

Some assets were reserved.

Unreserved

Fields

reserve_id: T::ReserveIdentifier

asset_id: T::AssetId

who: T::AccountId

amount: T::Balance

Some reserved assets were unreserved.

BurnedReserve

Fields

reserve_id: T::ReserveIdentifier

asset_id: T::AssetId

who: T::AccountId

amount: T::Balance

Some reserved assets were burned.

TransferredReserve

Fields

reserve_id: T::ReserveIdentifier

asset_id: T::AssetId

from: T::AccountId

to: T::AccountId

amount: T::Balance

Some reserved assets were transferred.

Burned

Fields

asset_id: T::AssetId

owner: T::AccountId

balance: T::Balance

Some assets were destroyed.

TeamChanged

Fields

asset_id: T::AssetId

issuer: T::AccountId

admin: T::AccountId

freezer: T::AccountId

The management team changed.

OwnerChanged

Fields

asset_id: T::AssetId

owner: T::AccountId

The owner changed.

Frozen

Fields

asset_id: T::AssetId

who: T::AccountId

Some account who was frozen.

Thawed

Fields

asset_id: T::AssetId

who: T::AccountId

Some account who was thawed.

AssetFrozen

Fields

asset_id: T::AssetId

Some asset asset_id was frozen.

AssetThawed

Fields

asset_id: T::AssetId

Some asset asset_id was thawed.

AccountsDestroyed

Fields

asset_id: T::AssetId

accounts_destroyed: u32

accounts_remaining: u32

Accounts were destroyed for given asset.

ApprovalsDestroyed

Fields

asset_id: T::AssetId

approvals_destroyed: u32

approvals_remaining: u32

Approvals were destroyed for given asset.

DestructionStarted

Fields

asset_id: T::AssetId

An asset class is in the process of being destroyed.

Destroyed

Fields

asset_id: T::AssetId

An asset class was destroyed. An asset class was destroyed.

ForceCreated

Fields

asset_id: T::AssetId

owner: T::AccountId

Some asset class was force-created.

MetadataSet

Fields

asset_id: T::AssetId

name: Vec<u8>

symbol: Vec<u8>

decimals: u8

is_frozen: bool

New metadata has been set for an asset.

MetadataCleared

Fields

asset_id: T::AssetId

Metadata has been cleared for an asset.

ApprovedTransfer

Fields

asset_id: T::AssetId

source: T::AccountId

delegate: T::AccountId

amount: T::Balance

(Additional) funds have been approved for transfer to a destination account.

ApprovalCancelled

Fields

asset_id: T::AssetId

owner: T::AccountId

delegate: T::AccountId

An approval for account delegate was cancelled by owner.

TransferredApproved

Fields

asset_id: T::AssetId

owner: T::AccountId

delegate: T::AccountId

destination: T::AccountId

amount: T::Balance

An amount was transferred in its entirety from owner to destination by the approved delegate.

AssetStatusChanged

Fields

asset_id: T::AssetId

An asset has had its attributes changed by the Force origin.