Struct pallet_mapped_assets::pallet::Pallet

pub struct Pallet<T, I = ()>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn adjust_extra(
id: T::AssetId,
who: impl Borrow<T::AccountId>
) -> Option<ExtraMutator<T, I>>

Return the extra “sid-car” data for id/who, or None if the account doesn’t exist.

pub fn balance(id: T::AssetId, who: impl Borrow<T::AccountId>) -> T::Balance

Get the asset id balance of who, or zero if the asset-account doesn’t exist.

pub fn maybe_balance(
id: T::AssetId,
who: impl Borrow<T::AccountId>
) -> Option<T::Balance>

Get the asset id balance of who if the asset-account exists.

pub fn reserved_balance(
id: T::AssetId,
who: impl Borrow<T::AccountId>
) -> T::Balance

Get the asset id reserved balance of who, or zero if the asset-account doesn’t exist.

pub fn maybe_reserved_balance(
id: T::AssetId,
who: impl Borrow<T::AccountId>
) -> Option<T::Balance>

Get the asset id reserved balance of who if the asset-account exists.

pub fn total_supply(id: T::AssetId) -> T::Balance

Get the total supply of an asset id.

pub fn maybe_total_supply(id: T::AssetId) -> Option<T::Balance>

Get the total supply of an asset id if the asset exists.

pub fn total_reserved_supply(id: T::AssetId) -> T::Balance

Get the total reserved supply of an asset id.

pub fn maybe_total_reserved_supply(id: T::AssetId) -> Option<T::Balance>

Get the total reversed supply of an asset id if the asset exists.

pub fn reserved_balance_named(
id: &T::ReserveIdentifier,
asset_id: &T::AssetId,
who: &T::AccountId
) -> T::Balance

pub fn has_named_reserve(
id: &T::ReserveIdentifier,
asset_id: &T::AssetId,
who: &T::AccountId
) -> bool

pub fn reserve_named(
id: &T::ReserveIdentifier,
asset_id: T::AssetId,
who: &T::AccountId,
amount: T::Balance,
maybe_check_admin: Option<T::AccountId>
) -> DispatchResult

Move value from the free balance from who to a named reserve balance.

Is a no-op if value to be reserved is zero.

pub fn unreserve_named(
id: &T::ReserveIdentifier,
asset_id: T::AssetId,
who: &T::AccountId,
maybe_check_admin: Option<T::AccountId>
) -> DispatchResult

pub fn burn_named_reserve(
id: &T::ReserveIdentifier,
asset_id: T::AssetId,
who: &T::AccountId,
maybe_check_admin: Option<T::AccountId>
) -> DispatchResult

pub fn transfer_named_reserve(
id: &T::ReserveIdentifier,
asset_id: T::AssetId,
from: &T::AccountId,
to: &T::AccountId,
maybe_check_admin: Option<T::AccountId>
) -> DispatchResult

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn create(
origin: OriginFor<T>,
id: T::AssetId,
admin: <<T as Config>::Lookup as StaticLookup>::Source,
min_balance: T::Balance
) -> DispatchResult

Issue a new class of fungible assets from a public origin.

This new asset class has no assets initially and its owner is the origin.

The origin must conform to the configured CreateOrigin and have sufficient funds free.

Funds of sender are reserved by AssetDeposit.

Parameters:

id: The identifier of the new asset. This must not be currently in use to identify an existing asset.

admin: The admin of this class of assets. The admin is the initial address of each member of the asset class’s admin team.

min_balance: The minimum balance of this new asset that any single account must have. If an account’s balance is reduced below this, then it collapses to zero.

Emits Created event when successful.

Weight: O(1)

pub fn force_create(
origin: OriginFor<T>,
id: T::AssetId,
owner: <<T as Config>::Lookup as StaticLookup>::Source,
is_sufficient: bool,
min_balance: T::Balance
) -> DispatchResult

Issue a new class of fungible assets from a privileged origin.

This new asset class has no assets initially.

The origin must conform to ForceOrigin.

Unlike create, no funds are reserved.

id: The identifier of the new asset. This must not be currently in use to identify an existing asset.

owner: The owner of this class of assets. The owner has full superuser permissions over this asset, but may later change and configure the permissions using transfer_ownership and set_team.

min_balance: The minimum balance of this new asset that any single account must have. If an account’s balance is reduced below this, then it collapses to zero.

Emits ForceCreated event when successful.

Weight: O(1)

pub fn start_destroy(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResult

Destroy a class of fungible assets.

The origin must conform to ForceOrigin or must be Signed and the sender must be the owner of the asset id.

id: The identifier of the asset to be destroyed. This must identify an existing asset.

Emits Destroyed event when successful.

NOTE: It can be helpful to first freeze an asset before destroying it so that you can provide accurate witness information and prevent users from manipulating state in a way that can make it harder to destroy.

Weight: O(c + p + a) where:

c = (witness.accounts - witness.sufficients)

s = witness.sufficients

a = witness.approvals Start the process of destroying a fungible asset class.

start_destroy is the first in a series of extrinsics that should be called, to allow destruction of an asset class.

The origin must conform to ForceOrigin or must be Signed by the asset’s owner.

id: The identifier of the asset to be destroyed. This must identify an existing asset.

The asset class must be frozen before calling start_destroy.

pub fn destroy_accounts(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResultWithPostInfo

Destroy all accounts associated with a given asset.

destroy_accounts should only be called after start_destroy has been called, and the asset is in a Destroying state.

Due to weight restrictions, this function may need to be called multiple times to fully destroy all accounts. It will destroy RemoveItemsLimit accounts at a time.

id: The identifier of the asset to be destroyed. This must identify an existing asset.

Each call emits the Event::DestroyedAccounts event.

pub fn destroy_approvals(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResultWithPostInfo

Destroy all approvals associated with a given asset up to the max (T::RemoveItemsLimit).

destroy_approvals should only be called after start_destroy has been called, and the asset is in a Destroying state.

Due to weight restrictions, this function may need to be called multiple times to fully destroy all approvals. It will destroy RemoveItemsLimit approvals at a time.

id: The identifier of the asset to be destroyed. This must identify an existing asset.

Each call emits the Event::DestroyedApprovals event.

pub fn finish_destroy(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResult

Complete destroying asset and unreserve currency.

finish_destroy should only be called after start_destroy has been called, and the asset is in a Destroying state. All accounts or approvals should be destroyed before hand.

id: The identifier of the asset to be destroyed. This must identify an existing asset.

Each successful call emits the Event::Destroyed event.

pub fn mint(
origin: OriginFor<T>,
id: T::AssetIdParameter,
beneficiary: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Mint assets of a particular class.

The origin must be Signed and the sender must be the Issuer of the asset id.

id: The identifier of the asset to have some amount minted.

beneficiary: The account to be credited with the minted assets.

amount: The amount of the asset to be minted.

Emits Issued event when successful.

Weight: O(1) Modes: Pre-existing balance of beneficiary; Account pre-existence of beneficiary.

pub fn burn(
origin: OriginFor<T>,
id: T::AssetIdParameter,
who: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Reduce the balance of who by as much as possible up to amount assets of id.

Origin must be Signed and the sender should be the Manager of the asset id.

Bails with NoAccount if the who is already dead.

id: The identifier of the asset to have some amount burned.

who: The account to be debited from.

amount: The maximum amount by which who’s balance should be reduced.

Emits Burned with the actual amount burned. If this takes the balance to below the minimum for the asset, then the amount burned is increased to take it to zero.

Weight: O(1) Modes: Post-existence of who; Pre & post Zombie-status of who.

pub fn transfer(
origin: OriginFor<T>,
id: T::AssetIdParameter,
target: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Move some assets from the sender account to another.

Origin must be Signed.

id: The identifier of the asset to have some amount transferred.

target: The account to be credited.

amount: The amount by which the sender’s balance of assets should be reduced and target’s balance increased. The amount actually transferred may be slightly greater in the case that the transfer would otherwise take the sender balance above zero but below the minimum balance. Must be greater than zero.

Emits Transferred with the actual amount transferred. If this takes the source balance to below the minimum for the asset, then the amount transferred is increased to take it to zero.

Weight: O(1) Modes: Pre-existence of target; Post-existence of sender; Account pre-existence of target.

pub fn transfer_keep_alive(
origin: OriginFor<T>,
id: T::AssetIdParameter,
target: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Move some assets from the sender account to another, keeping the sender account alive.

Origin must be Signed.

id: The identifier of the asset to have some amount transferred.

target: The account to be credited.

amount: The amount by which the sender’s balance of assets should be reduced and target’s balance increased. The amount actually transferred may be slightly greater in the case that the transfer would otherwise take the sender balance above zero but below the minimum balance. Must be greater than zero.

Emits Transferred with the actual amount transferred. If this takes the source balance to below the minimum for the asset, then the amount transferred is increased to take it to zero.

Weight: O(1) Modes: Pre-existence of target; Post-existence of sender; Account pre-existence of target.

pub fn force_transfer(
origin: OriginFor<T>,
id: T::AssetIdParameter,
source: <<T as Config>::Lookup as StaticLookup>::Source,
dest: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Move some assets from one account to another.

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier of the asset to have some amount transferred.

source: The account to be debited.

dest: The account to be credited.

amount: The amount by which the source’s balance of assets should be reduced and dest’s balance increased. The amount actually transferred may be slightly greater in the case that the transfer would otherwise take the source balance above zero but below the minimum balance. Must be greater than zero.

Emits Transferred with the actual amount transferred. If this takes the source balance to below the minimum for the asset, then the amount transferred is increased to take it to zero.

Weight: O(1) Modes: Pre-existence of dest; Post-existence of source; Account pre-existence of dest.

pub fn freeze(
origin: OriginFor<T>,
id: T::AssetIdParameter,
who: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Disallow further unprivileged transfers from an account.

Origin must be Signed and the sender should be the Freezer of the asset id.

id: The identifier of the asset to be frozen.

who: The account to be frozen.

Emits Frozen.

Weight: O(1)

pub fn thaw(
origin: OriginFor<T>,
id: T::AssetIdParameter,
who: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Allow unprivileged transfers from an account again.

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier of the asset to be frozen.

who: The account to be unfrozen.

Emits Thawed.

Weight: O(1)

pub fn freeze_asset(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResult

Disallow further unprivileged transfers for the asset class.

Origin must be Signed and the sender should be the Freezer of the asset id.

id: The identifier of the asset to be frozen.

Emits Frozen.

Weight: O(1)

pub fn thaw_asset(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResult

Allow unprivileged transfers for the asset again.

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier of the asset to be thawed.

Emits Thawed.

Weight: O(1)

pub fn transfer_ownership(
origin: OriginFor<T>,
id: T::AssetIdParameter,
owner: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Change the Owner of an asset.

Origin must be Signed and the sender should be the Owner of the asset id.

id: The identifier of the asset.

owner: The new Owner of this asset.

Emits OwnerChanged.

Weight: O(1)

pub fn set_team(
origin: OriginFor<T>,
id: T::AssetIdParameter,
issuer: <<T as Config>::Lookup as StaticLookup>::Source,
admin: <<T as Config>::Lookup as StaticLookup>::Source,
freezer: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Change the Issuer, Admin and Freezer of an asset.

Origin must be Signed and the sender should be the Owner of the asset id.

id: The identifier of the asset to be frozen.

issuer: The new Issuer of this asset.

admin: The new Admin of this asset.

freezer: The new Freezer of this asset.

Emits TeamChanged.

Weight: O(1)

pub fn set_metadata(
origin: OriginFor<T>,
id: T::AssetIdParameter,
name: Vec<u8>,
symbol: Vec<u8>,
decimals: u8
) -> DispatchResult

Set the metadata for an asset.

Origin must be Signed and the sender should be the Owner of the asset id.

Funds of sender are reserved according to the formula: MetadataDepositBase + MetadataDepositPerByte * (name.len + symbol.len) taking into account any already reserved funds.

id: The identifier of the asset to update.

name: The user friendly name of this asset. Limited in length by StringLimit.

symbol: The exchange symbol for this asset. Limited in length by StringLimit.

decimals: The number of decimals this asset uses to represent one unit.

Emits MetadataSet.

Weight: O(1)

pub fn clear_metadata(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResult

Clear the metadata for an asset.

Origin must be Signed and the sender should be the Owner of the asset id.

Any deposit is freed for the asset owner.

id: The identifier of the asset to clear.

Emits MetadataCleared.

Weight: O(1)

pub fn force_set_metadata(
origin: OriginFor<T>,
id: T::AssetIdParameter,
name: Vec<u8>,
symbol: Vec<u8>,
decimals: u8,
is_frozen: bool
) -> DispatchResult

Force the metadata for an asset to some value.

Origin must be ForceOrigin.

Any deposit is left alone.

id: The identifier of the asset to update.

name: The user friendly name of this asset. Limited in length by StringLimit.

symbol: The exchange symbol for this asset. Limited in length by StringLimit.

decimals: The number of decimals this asset uses to represent one unit.

Emits MetadataSet.

Weight: O(N + S) where N and S are the length of the name and symbol respectively.

pub fn force_clear_metadata(
origin: OriginFor<T>,
id: T::AssetIdParameter
) -> DispatchResult

Clear the metadata for an asset.

Origin must be ForceOrigin.

Any deposit is returned.

id: The identifier of the asset to clear.

Emits MetadataCleared.

Weight: O(1)

pub fn force_asset_status(
origin: OriginFor<T>,
id: T::AssetIdParameter,
owner: <<T as Config>::Lookup as StaticLookup>::Source,
issuer: <<T as Config>::Lookup as StaticLookup>::Source,
admin: <<T as Config>::Lookup as StaticLookup>::Source,
freezer: <<T as Config>::Lookup as StaticLookup>::Source,
min_balance: T::Balance,
is_sufficient: bool,
is_frozen: bool
) -> DispatchResult

Alter the attributes of a given asset.

Origin must be ForceOrigin.

id: The identifier of the asset.

owner: The new Owner of this asset.

issuer: The new Issuer of this asset.

admin: The new Admin of this asset.

freezer: The new Freezer of this asset.

min_balance: The minimum balance of this new asset that any single account must have. If an account’s balance is reduced below this, then it collapses to zero.

is_sufficient: Whether a non-zero balance of this asset is deposit of sufficient value to account for the state bloat associated with its balance storage. If set to true, then non-zero balances may be stored without a consumer reference (and thus an ED in the Balances pallet or whatever else is used to control user-account state growth).

is_frozen: Whether this asset class is frozen except for permissioned/admin instructions.

Emits AssetStatusChanged with the identity of the asset.

Weight: O(1)

pub fn approve_transfer(
origin: OriginFor<T>,
id: T::AssetIdParameter,
delegate: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Approve an amount of asset for transfer by a delegated third-party account.

Origin must be Signed.

Ensures that ApprovalDeposit worth of Currency is reserved from signing account for the purpose of holding the approval. If some non-zero amount of assets is already approved from signing account to delegate, then it is topped up or unreserved to meet the right value.

NOTE: The signing account does not need to own amount of assets at the point of making this call.

id: The identifier of the asset.

delegate: The account to delegate permission to transfer asset.

amount: The amount of asset that may be transferred by delegate. If there is already an approval in place, then this acts additively.

Emits ApprovedTransfer on success.

Weight: O(1)

pub fn cancel_approval(
origin: OriginFor<T>,
id: T::AssetIdParameter,
delegate: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Cancel all of some asset approved for delegated transfer by a third-party account.

Origin must be Signed and there must be an approval in place between signer and delegate.

Unreserves any deposit previously reserved by approve_transfer for the approval.

id: The identifier of the asset.

delegate: The account delegated permission to transfer asset.

Emits ApprovalCancelled on success.

Weight: O(1)

pub fn force_cancel_approval(
origin: OriginFor<T>,
id: T::AssetIdParameter,
owner: <<T as Config>::Lookup as StaticLookup>::Source,
delegate: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Cancel all of some asset approved for delegated transfer by a third-party account.

Origin must be either ForceOrigin or Signed origin with the signer being the Admin account of the asset id.

Unreserves any deposit previously reserved by approve_transfer for the approval.

id: The identifier of the asset.

delegate: The account delegated permission to transfer asset.

Emits ApprovalCancelled on success.

Weight: O(1)

pub fn transfer_approved(
origin: OriginFor<T>,
id: T::AssetIdParameter,
owner: <<T as Config>::Lookup as StaticLookup>::Source,
destination: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Transfer some asset balance from a previously delegated account to some third-party account.

Origin must be Signed and there must be an approval in place by the owner to the signer.

If the entire amount approved for transfer is transferred, then any deposit previously reserved by approve_transfer is unreserved.

id: The identifier of the asset.

owner: The account which previously approved for a transfer of at least amount and from which the asset balance will be withdrawn.

destination: The account to which the asset balance of amount will be transferred.

amount: The amount of assets to transfer.

Emits TransferredApproved on success.

Weight: O(1)

pub fn touch(origin: OriginFor<T>, id: T::AssetIdParameter) -> DispatchResult

Create an asset account for non-provider assets.

A deposit will be taken from the signer account.

origin: Must be Signed; the signer account must have sufficient funds for a deposit to be taken.

id: The identifier of the asset for the account to be created.

Emits Touched event when successful.

pub fn refund(
origin: OriginFor<T>,
id: T::AssetIdParameter,
allow_burn: bool
) -> DispatchResult

Return the deposit (if any) of an asset account.

The origin must be Signed.

id: The identifier of the asset for the account to be created.

allow_burn: If true then assets may be destroyed in order to complete the refund.

Emits Refunded event when successful.

pub fn reserve(
origin: OriginFor<T>,
id: T::ReserveIdentifier,
asset_id: T::AssetId,
who: <<T as Config>::Lookup as StaticLookup>::Source,
amount: T::Balance
) -> DispatchResult

Reserve some assets on an account for a specific asset.

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier to be used for the reserve

asset_id: The identifier of the asset.

who: The account on which the assets are to be reserved.

amount: The amount to reserve who’s free balance is decreased by amount and the reserve balance increased by amount. The reserve balance of the asset with asset_id is increased by amount. A named reserve with id is added to reserves.

Emits Reserved with the reserved amount.

Weight: O(1)

pub fn unreserve(
origin: OriginFor<T>,
id: T::ReserveIdentifier,
asset_id: T::AssetId,
who: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Unreserve assets reserved under id from an account

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier to be used for the reserve

asset_id: The identifier of the asset.

who: The account on which the assets are to be unreserved. who’s free balance is increased by amount and the reserve balance decreased by amount. The reserve balance of the asset with asset_id is decreased by amount. The named reserve with id is removed.

Emits Uneserved with the reserved amount.

Weight: O(1)

pub fn burn_reserve(
origin: OriginFor<T>,
id: T::ReserveIdentifier,
asset_id: T::AssetId,
who: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Burn assets reserved under id from an account

Origin must be Signed and the sender should be the Admin of the asset id.

id: The identifier to be used for the reserve

asset_id: The identifier of the asset.

who: The account on which the reserved assets are to be burned. who’s reserve balance is decreased by amount. The reserve balance of the asset with asset_id is decreased by amount and the supply is decreased by amount. The named reserve with id is removed.

Emits BurnedReserve with the reserved amount.

Weight: O(1)

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn reserves<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> BoundedVec<ReserveData<T::ReserveIdentifier, T::Balance>, T::MaxReserves>where
KArg1: EncodeLike<T::AssetId>,
KArg2: EncodeLike<T::AccountId>,

Named reserves on account balances for a specific asset.