Trait pallet_mapped_assets::pallet::Config

pub trait Config<I: 'static = ()>: Config {

    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen + TypeInfo;
    type RemoveItemsLimit: Get<u32>;
    type AssetId: Member + Parameter + Default + Copy + HasCompact + MaybeSerializeDeserialize + MaxEncodedLen + TypeInfo;
    type AssetIdParameter: Parameter + Copy + From<Self::AssetId> + Into<Self::AssetId> + MaxEncodedLen;
    type Currency: ReservableCurrency<Self::AccountId>;
    type CreateOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, Self::AssetId, Success = Self::AccountId>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type AssetDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type AssetAccountDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type MetadataDepositBase: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type MetadataDepositPerByte: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type ApprovalDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type StringLimit: Get<u32>;
    type MaxReserves: Get<u32>;
    type Freezer: FrozenBalance<Self::AssetId, Self::AccountId, Self::Balance>;
    type Extra: Member + Parameter + Default + MaxEncodedLen;
    type CallbackHandle: AssetsCallback<Self::AssetId, Self::AccountId>;
    type WeightInfo: WeightInfo;
    type ReserveIdentifier: Parameter + Member + MaxEncodedLen + Ord + Copy;

}

The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen + TypeInfo

The units in which we record balances.

type RemoveItemsLimit: Get<u32>

Max number of items to destroy per destroy_accounts and destroy_approvals call.

Must be configured to result in a weight that makes each call fit in a block.

type AssetId: Member + Parameter + Default + Copy + HasCompact + MaybeSerializeDeserialize + MaxEncodedLen + TypeInfo

Identifier for the class of asset.

type AssetIdParameter: Parameter + Copy + From<Self::AssetId> + Into<Self::AssetId> + MaxEncodedLen

Wrapper around Self::AssetId to use in dispatchable call signatures. Allows the use of compact encoding in instances of the pallet, which will prevent breaking changes resulting from the removal of HasCompact from Self::AssetId.

This type includes the From<Self::AssetId> bound, since tightly coupled pallets may want to convert an AssetId into a parameter for calling dispatchable functions directly.

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism.

type CreateOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, Self::AssetId, Success = Self::AccountId>

Standard asset class creation is only allowed if the origin attempting it and the asset class are in this set.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly create or destroy an asset or otherwise alter privileged attributes.

type AssetDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The basic amount of funds that must be reserved for an asset.

type AssetAccountDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The amount of funds that must be reserved for a non-provider asset account to be maintained.

type MetadataDepositBase: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The basic amount of funds that must be reserved when adding metadata to your asset.

type MetadataDepositPerByte: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The additional funds that must be reserved for the number of bytes you store in your metadata.

type ApprovalDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The amount of funds that must be reserved when creating a new approval.

type StringLimit: Get<u32>

The maximum length of a name or symbol stored on-chain.

type MaxReserves: Get<u32>

The maximum number of named reserves that can exist on an account.

type Freezer: FrozenBalance<Self::AssetId, Self::AccountId, Self::Balance>

A hook to allow a per-asset, per-account minimum balance to be enforced. This must be respected in all permissionless operations.

type Extra: Member + Parameter + Default + MaxEncodedLen

Additional data to be stored with an account’s asset balance.

type CallbackHandle: AssetsCallback<Self::AssetId, Self::AccountId>

Callback methods for asset state change (e.g. asset created or destroyed)

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type ReserveIdentifier: Parameter + Member + MaxEncodedLen + Ord + Copy

The id type for named reserves.