Module pallet_mapped_assets::types

Various basic types for use in the assets pallet.

Structs

Approval

Data concerning an approval.

AssetAccount

AssetDetails

AssetMetadata

BalanceToAssetBalance

Converts a balance value into an asset balance based on the ratio between the fungible’s minimum balance and the minimum asset balance.

ReserveData

Store named reserved balance.

Enums

ConversionError

Possible errors when converting between external and asset balances.

ExistenceReason

Traits

FrozenBalance

Trait for allowing a minimum balance on the account to be specified, beyond the minimum_balance of the asset. This is additive - the minimum_balance of the asset must be met and then anything here in addition.